requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
cypari==2.5.5
db-dtypes==1.4.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.13
fastjsonschema==2.21.1
ffmpeg==1.4
ffmpeg-python==0.2.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
future==1.0.0
fxrays==1.3.6
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.35.0
ipywidgets==8.1.6
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-pygments==0.3.0
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
kneed==0.8.5
knot-floer-homology==1.2.2
llvmlite==0.44.0
low-index==1.2.1
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.3.7
plink==2.4.6
plotly==5.20.0
prettytable==3.16.0
prometheus-client==0.21.1
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
pyx==0.16
pyyaml==6.0.2
pyzmq==26.4.0
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.4.0
scipy==1.15.2
seaborn==0.13.2
send2trash==1.8.3
setuptools==78.1.0
six==1.17.0
snappy==3.2
snappy-manifolds==1.2.1
sniffio==1.3.1
soupsieve==2.6
spherogram==2.3
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tkinter-gl==1.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
umap-learn==0.5.6
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
yarl==1.20.1
